Refactor Projects To be more modular
