+++ src/components/layout/projects/Professional.jsx
@@ -1,66 +1,12 @@
-import HeroImage from "/images/Hero-Image-Mobile.jpg";
-
-const professionalProjects = [
-  {
-    title: "The Green Pig Pub (Freelance)",
-    content:
-      "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
-    image: {
-      HeroImage,
-      alt: "Image of The Green Pig Pub's Website",
-    },
-    url: "https://lordsahvith.github.io/green-pig-pub/index.html",
-  },
-  {
-    title: "Groupe Dynamite Inc",
-    content:
-      "(Dynamite & Garage) Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
-    image: {
-      HeroImage,
-      alt: "Image of Yu-Gi-Oh Card Colletor Project",
-    },
-    url: "https://www.dynamiteclothing.com/us/",
-  },
-  {
-    title: "Cricut",
-    content:
-      "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
-    image: {
-      HeroImage,
-      alt: "Image of Yu-Gi-Oh Card Colletor Project",
-    },
-    url: "https://cricut.com/en-us/",
-  },
-  {
-    title: "Cricut Heat Guide",
-    content:
-      "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
-    image: {
-      HeroImage,
-      alt: "Image of Yu-Gi-Oh Card Colletor Project",
-    },
-    url: "https://cricut.com/en-us/heatguide",
-  },
-];
+import ProjectTemplate from "./ProjectTemplate";
+import { professionalProjects } from "./data/projects";
 
 function ProfessionalProjects() {
   return (
     <div>
       <h3>Professional</h3>
 
-      {professionalProjects.map((project) => {
-        return (
-          <>
-            <div>
-              <h4>{project.title}</h4>
-              <p>{project.content}</p>
-            </div>
-            <div>
-              <img src={project.image.HeroImage} alt={project.image.alt} />
-            </div>
-          </>
-        );
-      })}
+      <ProjectTemplate projects={professionalProjects} />
     </div>
   );
 }
