+++ src/components/layout/projects/data/projects.js
@@ -1,0 +1,58 @@
+import HeroImage from "/images/Hero-Image-Mobile.jpg";
+
+const personalProjects = [
+  {
+    title: "Yu-Gi-Oh Card Collector",
+    content:
+      "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
+    image: {
+      HeroImage,
+      alt: "Image of Yu-Gi-Oh Card Colletor Project",
+    },
+  },
+];
+
+const professionalProjects = [
+  {
+    title: "The Green Pig Pub (Freelance)",
+    content:
+      "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
+    image: {
+      HeroImage,
+      alt: "Image of The Green Pig Pub's Website",
+    },
+    url: "https://lordsahvith.github.io/green-pig-pub/index.html",
+  },
+  {
+    title: "Groupe Dynamite Inc",
+    content:
+      "(Dynamite & Garage) Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
+    image: {
+      HeroImage,
+      alt: "Image of Groupe Dynamite Website",
+    },
+    url: "https://www.dynamiteclothing.com/us/",
+  },
+  {
+    title: "Cricut",
+    content:
+      "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
+    image: {
+      HeroImage,
+      alt: "Image Cricut Website",
+    },
+    url: "https://cricut.com/en-us/",
+  },
+  {
+    title: "Cricut Heat Guide",
+    content:
+      "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
+    image: {
+      HeroImage,
+      alt: "Image Cricut Heat Guide",
+    },
+    url: "https://cricut.com/en-us/heatguide",
+  },
+];
+
+export { personalProjects, professionalProjects };
